Move reference files to a proper place
