+++ privacy-config/privacy-config-impl/src/test/resources/reference_tests/trackerallowlist/tracker_allowlist_reference.json
